Add basic FTM timestampts for transmitted packets to tx_intf
Adds a new clock-driven counter to the XPU. It starts at zero and is incremented every clock tick.
The counter is exposed as an output of the XPU ip, and is connected to a matching input on the tx_intf
ip.
It is used in the tx_intf module to create timestamps for packet transmission and acknowledgement receival timestamps
to implement the fine time measurement (FTM) protocol.
The tx_intf module listens to openofdm_tx signals to detect start of transmission, and xpu to detect receival of ACK,
and takes timestamps for both. These timestamps are exposed to software via new AXI registers, using the existing status information
queues implementation.